backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
SQLAlchemy==2.0.35
pydantic==2.9.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
PyMySQL==1.1.1
requests==2.32.3
pycryptodome==3.21.0
python-dotenv==1.0.1

